webapp/backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
scikit-learn
numpy
pandas
joblib
python-dotenv
pydantic
